refactor: rename `bin` to `QraftCommand`
