+++ packages/cli/src/lib/QraftCommandPlugin.ts
@@ -1,4 +1,4 @@
-import { QraftCommand } from '../bin.js';
+import { QraftCommand } from './QraftCommand.js';
 
 export interface QraftCommandPlugin {
   setupCommand(command: QraftCommand): void;
